Moved mock objects for Autoconfig to a separate file to facilitate reuse.
